life transforming messagese added in playlist and playlist added in seperate js file
