+++ index.html
@@ -72,6 +72,7 @@
       </div>
       <div id="playlistsContainer"></div>
     </div>
+    <script src="./js/playlist.js"></script>
     <script src="./js/script.js"></script>
     <script src="./js/nav.js"></script>
   </body>
