+++ js/playlist.js
@@ -1,0 +1,300 @@
+const playlists = [
+  {
+    name: "Bhajans",
+    tracks: [
+      {
+        title: "Racha Hai Srishti Ko Jis Prabhu Ne",
+        url: "./bhajan/mp3/RachaHaiSrishtiKoJisPrabhuNe.mp3",
+      },
+      {
+        title: "Ram Nam Ke Hire Moti",
+        url: "./bhajan/mp3/RamNamKeHireMoti.mp3",
+      },
+      {
+        title: "Mujhe Tumne Malik Bahut Diya Hai",
+        url: "./bhajan/mp3/MujheTumneMalikBahutDeDiyahai.mp3",
+      },
+      {
+        title: "Vishnu Sahasranamam",
+        url: "./bhajan/mp3/VishnuSahasranamam.mp3",
+      },
+      {
+        title: "Nar Sharir Anmol Hai Prani",
+        url: "./bhajan/mp3/NarSharirAnmolHaiPrani.mp3",
+      },
+      {
+        title: "Saare Thirath Dhaam Aapke",
+        url: "./bhajan/mp3/Saare-Thirath-Dhaam-Aapke.mp3",
+      },
+    ],
+  },
+  {
+    name: "Shiv Bhajans",
+    tracks: [
+      {
+        title: "Shiv Amritwani",
+        url: "./bhajan/shiv/Shiv-Amritwani.mp3",
+      },
+      {
+        title: "Karpur Gauram Karunavtaram",
+        url: "./bhajan/shiv/KarpurGauramKarunavtaram.mp3",
+      },
+      {
+        title: "Bhole Girijapati Hu Tamhari Sharan",
+        url: "./bhajan/shiv/BholeGirijapatihuTamhariSharan.mp3",
+      },
+      {
+        title: "Hey Shiv Pita Parmatma",
+        url: "./bhajan/shiv/HeyShivPitaParmatma.mp3",
+      },
+      {
+        title: "Ram Rachit Shambu Stuti",
+        url: "./bhajan/shiv/RamRachitShambuStuti.mp3",
+      },
+      {
+        title: "Satyam Shivam Sundaram",
+        url: "./bhajan/shiv/SatyamShivamSundaram.mp3",
+      },
+      {
+        title: "Lingaashtakam",
+        url: "./bhajan/shiv/Lingaashtakam.mp3",
+      },
+      {
+        title: "Shiv Dhun Om Namah Shivay",
+        url: "./bhajan/shiv/ShivDhunOmNamahShivay.mp3",
+      },
+      {
+        title: "Jagat Chatna Hu Anadi Ananta",
+        url: "./bhajan/shiv/JagatChatnaHuAnadiAnanta.mp3",
+      },
+      {
+        title: "Mahamrityunjay Mantra",
+        url: "./bhajan/shiv/MahamrityunjayMantra.mp3",
+      },
+      {
+        title: "Aisi Subah Na Aaye",
+        url: "./bhajan/shiv/Aisi-Subah-Na-Aaye.mp3",
+      },
+      {
+        title: "Mahadev Shankar Hai Jag Se Nirale",
+        url: "./bhajan/shiv/Mahadev-Shankar-Hai-Jag-Se-Nirale.mp3",
+      },
+      {
+        title: "Prabhu Mere Man Ko",
+        url: "./bhajan/shiv/Prabhu_Mere_Man_Ko.mp3",
+      },
+      {
+        title: "Tera Pal Pal Beeta Jaye",
+        url: "./bhajan/shiv/TERA-PAL-PAL-BEETA-JAYE.mp3",
+      },
+      {
+        title: "Top 10 Shiv Bhajans By Gulshan Kumar",
+        url: "./bhajan/shiv/Top-10-Best-Shiv-Bhajans-By-Gulshan-Kumar.mp3",
+      },
+    ],
+  },
+  {
+    name: "Vimal Playlist",
+    tracks: [
+      {
+        title: "Samb Sadashiv Meditation Version",
+        url: "./bhajan/vimal-playlist/1-Samb-Sadashiv-Meditation-Version.mp3",
+      },
+      {
+        title: "Samb Sadashiv Mantra 108",
+        url: "./bhajan/vimal-playlist/2-Samb-Sadashiv-Mantra-108.mp3",
+      },
+      {
+        title: "Samb Sadashiv Mantra Har Har Mahadev",
+        url: "./bhajan/vimal-playlist/3-Samb-Sadashiv-Mantra-Har-Har-Mahadev.mp3",
+      },
+      {
+        title: "Buddhist Sound Therapy",
+        url: "./bhajan/vimal-playlist/4-Buddhist-Sound-Therapy-for-Anxiety-and-Stress-Management.mp3",
+      },
+    ],
+  },
+  {
+    name: "Mata Ke Bhajans",
+    tracks: [
+      {
+        title: "Shri Durga Raksha Kavach",
+        url: "./bhajan/mp4/shri-durga-raksha-kavach.mp4",
+      },
+      {
+        title: "Jay Ambe Gauri Maiya Jay Shyaama Gauri",
+        url: "./bhajan/mata/Jay-Ambe-Gauri-Maiya-Jay-Shyaama-Gauri.mp3",
+      },
+      {
+        title: "Maa Maa Teri Mamta Se Mithi",
+        url: "./bhajan/mata/Maa-Maa-Teri-Mamta-Se-Mithi.mp3",
+      },
+      {
+        title: "Mann Mera Mandir",
+        url: "./bhajan/mata/MANN-MERA-MANDIR.mp3",
+      },
+      {
+        title: "Mere Naino Ki Pyas Bujha De",
+        url: "./bhajan/mata/Mere-Naino-Ki-Pyas-Bujha-De.mp3",
+      },
+      {
+        title: "Sar Ko Jhuka Lo Sherawali Ko Mana Lo",
+        url: "./bhajan/mata/Sar-Ko-Jhuka-Lo-Sherawali-Ko-Mana-Lo.mp3",
+      },
+      {
+        title: "Ayigiri Nandini Mahishasura Marddini Sthothram",
+        url: "./bhajan/mata/Ayigiri-Nandini-Mahishasura-Marddini-Sthothram.mp3",
+      },
+    ],
+  },
+  {
+    name: "Shri Ram & Hanuman Ji",
+    tracks: [
+      {
+        title: "Ram Amritwani",
+        url: "./bhajan/ram-hanuman/Ram_Amritwani_By_Anuradha_Paudwal.mp3",
+      },
+      {
+        title: "Mangal Bhavan Amangal Hari",
+        url: "./bhajan/ram-hanuman/Mangal_Bhavan_Amangal_Hari.mp3",
+      },
+      {
+        title: "Shri Hanuman Chalisa",
+        url: "./bhajan/ram-hanuman/Shri-Hanuman-Chalisa.mp3",
+      },
+      {
+        title: "Veer Hanumana Aati Balwana",
+        url: "./bhajan/ram-hanuman/Veer-Hanumana-Aati-Balwana.mp3",
+      },
+      {
+        title: "Sundar Kand By Ravindra Jain",
+        url: "./bhajan/ram-hanuman/Sundar-Kand.mp3",
+      },
+    ],
+  },
+  {
+    name: "Radhe Krishna Bhajans",
+    tracks: [
+      {
+        title: "Radha Ki Kirpa",
+        url: "./bhajan/radhe-krishna/Radha-Ki-Kirpa.mp3",
+      },
+      {
+        title: "Radha Nam Sang Brij 84 Kosh Yatra",
+        url: "./bhajan/radhe-krishna/Radha-Nam-Sang-Brij-84-Kosh-Yatra.mp3",
+      },
+      {
+        title: "Radhe Albeli Sarkar",
+        url: "./bhajan/radhe-krishna/Radhe-Albeli-Sarkar-remix.mp3",
+      },
+      {
+        title: "Radhe Radhe Japa Karo",
+        url: "./bhajan/radhe-krishna/Radhe-Radhe-Japa-Karo.mp3",
+      },
+      {
+        title: "Kisori Kuch Aisa Intjam Ho Jaye",
+        url: "./bhajan/radhe-krishna/KISHORIKUCHAISAINTJAMHOJAYE.mp3",
+      },
+      {
+        title: "Vida He Murlidhar Vida",
+        url: "./bhajan/radhe-krishna/Vida_he_murlidhar_vida.mp3",
+      },
+      {
+        title: "Ye To Prem Ki Bat Hia Udho",
+        url: "./bhajan/radhe-krishna/Ye-To-Prem-Ki-Bat-Hin-Udho.mp3",
+      },
+      {
+        title: "Shri krishna bhajans",
+        url: "./bhajan/radhe-krishna/Shri_krishna_bhajans.mp3",
+      },
+
+      {
+        title: "Krishna Govind Gopal Madhav Hare",
+        url: "./bhajan/radhe-krishna/Krishna_govind_gopal_madhav_hare.mp3",
+      },
+      {
+        title: "Pakad Lo Hath Banvari Nahin To",
+        url: "./bhajan/radhe-krishna/PakadLoHathBanvariNahinToDoobJaenge.mp3",
+      },
+    ],
+  },
+  {
+    name: "Meditation Sounds",
+    tracks: [
+      {
+        title: "1 Hour Version  Brahmananda Swaroopa",
+        url: "./bhajan/meditation-sounds/1-Hour-Version-Brahmananda-Swaroopa.mp3",
+      },
+    ],
+  },
+  {
+    name: "Life Transforming Messages",
+    tracks: [
+      {
+        title: "Remind Yourself This One Thing Every Hour",
+        url: "./media/life-transforming-messages/Remind-Yourself-This-One-Thing-Every-Hour.mp4",
+      },
+      {
+        title: "Preparing to receive Sadhguru's grace",
+        url: "./media/life-transforming-messages/Preparing-to-receive-Sadhguru's-grace.mp4",
+      },
+      {
+        title: "The Power of “Aum Namah Shivaya”",
+        url: "./media/life-transforming-messages/The-Power-of-“Aum-Namah-Shivaya”.mp4",
+      },
+      {
+        title: "Universe will Bless You",
+        url: "./media/life-transforming-messages/Universe-will-Bless-You.mp4",
+      },
+      {
+        title: "A Guru’s Extreme Lesson",
+        url: "./media/life-transforming-messages/A-Guru’s-Extreme-Lesson.mp4",
+      },
+      {
+        title: "Focus Will Be Created Only on Desire",
+        url: "./media/life-transforming-messages/Focus-Will-Be-Created-Only-on-Desire.mp4",
+      },
+      {
+        title: "Do This Before You Take a Dip",
+        url: "./media/life-transforming-messages/Do-This-Before-You-Take-a-Dip.mp4",
+      },
+      {
+        title: "Do you imbibe or expand",
+        url: "./media/life-transforming-messages/Do-you-imbibe-or-expand.mp4",
+      },
+      {
+        title:
+          "Powerful Pratyahara Yoga! Learn to Sit Like This with Eyes Closed",
+        url: "./media/life-transforming-messages/Powerful-Pratyahara-Yoga!-Learn-to-Sit-Like-This-with-Eyes-Closed.mp4",
+      },
+      {
+        title: "Follow Your Dreams Or Obey Your Parents Wishes",
+        url: "./media/life-transforming-messages/Follow-Your-Dreams-Or-Obey-Your-Parents-Wishes.mp4",
+      },
+      {
+        title: "Don't argue just follow this and move on",
+        url: "./media/life-transforming-messages/Don't-argue-just-follow-this-and-move-on.mp4",
+      },
+      {
+        title: "This Is Why the Mind Gets Tired Quickly",
+        url: "./media/life-transforming-messages/This-Is-Why-the-Mind-Gets-Tired-Quickly.mp4",
+      },
+      {
+        title: "Vivek Oberoi Explain Savings",
+        url: "./media/life-transforming-messages/Vivek-Oberoi-Explain-Savings.mp4",
+      },
+      {
+        title: "Workout After Delivery",
+        url: "./media/life-transforming-messages/Workout-After-Delivery.mp4",
+      },
+      {
+        title: "Hanuman dand",
+        url: "./media/life-transforming-messages/Hanuman-dand.mp4",
+      },
+      {
+        title: "Workout Pushups",
+        url: "./media/life-transforming-messages/Workout-Pushups.mp4",
+      },
+    ],
+  },
+];
